unit test for authentication is upgraded based on comment.
