--- hosting/tests/__init__.py
@@ -1,5 +1,0 @@
-from hosting.tests.test_forms import *
-from hosting.tests.test_models import *
-from hosting.tests.test_views import *
-from hosting.tests.tests impot *
-from hosting.tests.auth import *
