--- hosting/tests/auth.py
@@ -1,259 +1,0 @@
-from django.conf import settings
-from django.contrib.auth.tokens import default_token_generator
-from django.core import mail
-from django.core.mail import EmailMultiAlternatives
-from django.core.urlresolvers import reverse, reverse_lazy
-from django.test import TestCase, Client
-from django.template.loader import render_to_string
-from django.utils.encoding import force_bytes
-from django.utils.http import urlsafe_base64_encode, urlsafe_base64_decode
-from django.utils.translation import ugettext_lazy as _
-
-from membership.models import CustomUser
-
-from model_mommy import mommy
-from hosting.forms import HostingUserLoginForm, HostingUserSignupForm
-from utils.forms import PasswordResetRequestForm, SetPasswordForm
-
-
-class RegistrationUnitTest(TestCase):
-    def setUp(self):
-        self.client = Client()
-        self.resp = self.client.post(
-                reverse('hosting:signup'), 
-                data={
-                        'name': 'test', 
-                        'email': 'test@example.com', 
-                        'password': 'secret', 
-                        'confirm_password': 'secret'
-                    }
-            )
-    def test_login_redirection(self):
-        new_user = CustomUser.objects.get(email='test@example.com')
-        activation_resp = self.client.get(
-                reverse(
-                        'hosting:validate', 
-                        kwargs={ 'validate_slug': new_user.validation_slug }
-                    )
-            )
-        
-        login_resp = self.client.login(email=new_user.email, password='secret')
-        self.assertEqual(login_resp, True)
-
-        redirect_resp = self.client.post(
-            reverse('hosting:login'),
-            data={
-                'email': new_user.email,
-                'password': 'secret'
-            }
-        )
-        
-        self.assertRedirects(redirect_resp, reverse('hosting:dashboard'))
-
-    def test_registration(self):        
-        self.assertRedirects(self.resp, reverse('hosting:signup-validate'))
-
-        new_user = CustomUser.objects.get(email='test@example.com')
-
-        self.failUnless(new_user.check_password('secret'))
-        self.assertEqual(new_user.email, 'test@example.com')
-
-        self.assertEqual(len(mail.outbox), 1)
-
-    def test_activation(self):        
-        new_user = CustomUser.objects.get(email='test@example.com')
-        resp = self.client.get(
-                reverse(
-                        'hosting:validate', 
-                        kwargs={
-                                'validate_slug': new_user.validation_slug
-                            }
-                    )
-            )
-        self.assertEqual(200, resp.status_code)
-
-    def test_reset_password(self):
-        new_user = CustomUser.objects.get(email='test@example.com')
-        context = {
-                'user': new_user,
-                'token': default_token_generator.make_token(new_user),
-                'uid': urlsafe_base64_encode(force_bytes(new_user.pk)),
-                'site_name': 'test_site',
-                'base_url': "http://localhost:8000/"
-            }
-        resp = self.client.post(
-                reverse('hosting:reset_password'),
-                data={
-                        'email': 'test@example.com', 
-                        'subject': _('Password Reset'), 
-                        'context': context, 
-                        'template_name': 'password_reset_email', 
-                        'template_path': 'hosting/emails/'
-                    }
-            )
-
-        self.assertEqual(len(mail.outbox), 2)
-        self.assertRedirects(resp, reverse('hosting:login'))
-        
-    def test_resend_activation_link(self):
-        new_user = CustomUser.objects.get(email='test@example.com')
-        context = {'base_url': "http://localhost:8000/",
-                'activation_link': reverse_lazy(
-                        'hosting:validate',
-                        kwargs={'validate_slug': new_user.validation_slug}
-                    ),
-                'dcl_text': 'Data Center Light',
-            }
-        resp = self.client.post(
-                reverse('hosting:resend_activation_link'),
-                data={
-                        'subject': 'Data Center Light Account Activation Test',
-                        'to': 'test@example.com',
-                        'context': context,
-                        'template_name': 'hosting/reset_password.html',
-                        'template_path': 'datacenterlight/emails/',
-                        'from_address': '(Data Center Light) Data Center Light Support <support@datacenterlight.ch>'
-                    }
-            )
-        self.assertEqual(len(mail.outbox), 1)
-        self.assertEqual(200, resp.status_code)
-
-
-class EmailTest(TestCase):
-    def setUp(self):
-        self.client = Client()
-        self.data = {
-                'name': 'test', 
-                'email': 'test@example.com', 
-                'password': 'secret', 
-                'confirm_password': 'secret'
-            }
-        self.resp = self.client.post(
-                reverse('hosting:signup'), 
-                data=self.data
-            )
-
-    def test_mail_receiver_sender(self):
-        email = mail.outbox[0]
-        self.assertEqual(email.to[0], 'test@example.com')
-        self.assertEqual(email.from_email, settings.DCL_SUPPORT_FROM_ADDRESS)
-
-    def test_mail_type_content_to_user(self):
-        new_user = CustomUser.objects.get(email='test@example.com')
-        template_path = 'datacenterlight/emails/'
-        template_name = 'user_activation'
-
-        email_data = {
-                'subject': '{dcl_text} {account_activation}'.format(
-                    dcl_text=settings.DCL_TEXT,
-                    account_activation=_('Account Activation')
-                ),
-                'from_address': settings.DCL_SUPPORT_FROM_ADDRESS,
-                'to': new_user.email,
-                'context': {'base_url': "http://localhost:8000/",
-                            'activation_link': reverse(
-                                'hosting:validate',
-                                kwargs={
-                                    'validate_slug': new_user.validation_slug}),
-                            'dcl_text': settings.DCL_TEXT
-                            },
-                'template_name': template_name,
-                'template_path': template_path
-            }
-
-        template_full_path = '%s%s' % (template_path, template_name)
-        text_content = render_to_string('%s.txt' % template_full_path, email_data['context'])
-        html_content = render_to_string('%s.html' % template_full_path, email_data['context'])
-
-        email = EmailMultiAlternatives(email_data['subject'], text_content)
-        email.attach_alternative(html_content, "text/html")
-
-        email.to = [email_data.get('to', 'info@ungleich.ch')]
-
-        email.send()
-
-        self.assertEqual(mail.outbox[1].body, text_content)
-        self.assertEqual(mail.outbox[1].alternatives[0][0], html_content)
-
-
-class UserLoginFormTest(TestCase):
-    def setUp(self):
-        password = 'user_password'
-        self.user = mommy.make('CustomUser', validated=1)
-        self.user.set_password(password)
-        self.user.save()
-        self.completed_data = {
-            'email': self.user.email,
-            'password': password
-        }
-        self.incorrect_data = {
-            'email': 'test',
-        }
-
-    def test_valid_form(self):
-        form = HostingUserLoginForm(data=self.completed_data)
-        self.assertTrue(form.is_valid())
-
-    def test_invalid_form(self):
-        form = HostingUserLoginForm(data=self.incorrect_data)
-        self.assertFalse(form.is_valid())
-
-
-class UserSignupFormTest(TestCase):
-    def setUp(self):
-        self.completed_data = {
-            'name': 'test name',
-            'email': 'test@ungleich.ch',
-            'password': 'test_password',
-            'confirm_password': 'test_password'
-        }
-
-        self.incorrect_data = {
-            'email': 'test',
-        }
-
-    def test_valid_form(self):
-        form = HostingUserSignupForm(data=self.completed_data)
-        self.assertTrue(form.is_valid())
-
-    def test_invalid_form(self):
-        form = HostingUserSignupForm(data=self.incorrect_data)
-        self.assertFalse(form.is_valid())
-
-
-class PasswordResetRequestFormTest(TestCase):
-    def setUp(self):
-        self.user = mommy.make('CustomUser')
-        self.completed_data = {
-            'email': self.user.email,
-        }
-        self.incorrect_data = {
-            'email': 'test',
-        }
-
-    def test_valid_form(self):
-        form = PasswordResetRequestForm(data=self.completed_data)
-        self.assertTrue(form.is_valid())
-
-    def test_invalid_form(self):
-        form = PasswordResetRequestForm(data=self.incorrect_data)
-        self.assertFalse(form.is_valid())
-
-
-class SetPasswordFormTest(TestCase):
-    def setUp(self):
-        self.completed_data = {
-            'new_password1': 'new_password',
-            'new_password2': 'new_password',
-        }
-        self.incorrect_data = {
-            'email': 'test',
-        }
-
-    def test_valid_form(self):
-        form = SetPasswordForm(data=self.completed_data)
-        self.assertTrue(form.is_valid())
-
-    def test_invalid_form(self):
-        form = SetPasswordForm(data=self.incorrect_data)
-        self.assertFalse(form.is_valid())
